Remove `inf` from the list of keywords
`inf` is not a real keyword in Protobuf; it's part of textformat, which
gloms onto Protobuf in a slightly odd way. The same is true of `nan`,
`true`, `false` (and `True`, `False`, `T`, and `F`), which are not
listed as keywords. It seems best to keep them separate.

Ideally, the production for `option` should simply say "the value is any
textformat value", but that does not match what `protoc` does today.
I'd like for us to fix that in the fullness of time, but that seems out
of scope for this change.