Protoc allows "-nan" and is lenient for "inf" and "nan" inside message literals (#37)
This requires adding `nan` to the list of keywords, so we can model the
special float value `-nan` in the grammar.

I also made some minor simplifications in here by combining
`UintLiteral` and `IntLiteral` into a single `IntLiteral` production
(which can be optionally signed).

This also contains a correction to the set of identifiers that can be
used for boolean values in the text format. The previous text indicated
that `T` and `F` could be used, but that's incorrect: it's lower-case
`t` and `f` that can be used.