Clarify that no whitespace or comments are allowed before byte-order mark (#70)
In the "Syntax" section of the spec, whitespace and comments are ignored
and effectively discarded by the lexer. (This obviously isn't _quite_
true since comments need to be retained in some form in order to produce
source code info in a descriptor. But for the purposes of the EBNF
grammar presented, they are ignored.) However, there is one exception to
this that was not clearly called out: the optional `byte_order_mark` at
the beginning of the file must be the first three bytes of the file --
preceding whitespace or comments are not allowed. (While likely
self-evident to those familiar with Unicode text encoding, it seemed
useful to clarify for any other readers.)